[Refactor] delete PaymentCreator
